requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.1
Werkzeug==3.0.1 